requirements.txt
Flask==3.0.0
flask-cors==4.0.0
flask-login==0.6.3
flask-bcrypt==1.0.1
python-dotenv==1.0.0
requests==2.31.0
qrcode==7.4.2
Pillow==10.1.0
urllib3==2.1.0
